val_backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2
django-cors-headers==3.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3

